cloud-funcs/storing-api/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore==2.3.3